refactor(framework) Move generic inflatable utils out of gRPC utils (#5496)
