+++ framework/py/flwr/server/grid/grpc_grid.py
@@ -30,9 +30,11 @@ from flwr.common.constant import (
 from flwr.common.grpc import create_channel, on_channel_state_change
 from flwr.common.inflatable import get_all_nested_objects
 from flwr.common.inflatable_grpc_utils import (
-    inflate_object_from_contents,
     make_pull_object_fn_grpc,
     make_push_object_fn_grpc,
+)
+from flwr.common.inflatable_utils import (
+    inflate_object_from_contents,
     pull_objects,
     push_objects,
 )
